domain changes. add new domain service project.

+++ src/OnlineAssesment.Domain/Examination/Examination.cs
@@ -23,7 +23,7 @@ namespace OnlineAssesment.Domain
 
         public virtual ICollection<ExaminationQuestion> Questions { get; set; }
 
-        public int TotalScore {
+        public float TotalScore {
             get {
                 return Questions.Sum(q => q.Score);
             }
